requirements.txt
pyircsdk~=1.5.2
SQLAlchemy~=2.0.29
mjml~=0.11.0
pyotp~=2.9.0
python-dotenv~=1.0.1
jinja2~=3.1.3
